# docs/assets/api-spec.yml
openapi: 3.0.0
info:
  title: FFmpeg Docker API
  version: 0.1.0
  description: Common FFMPEG fucntions from a RESTful API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  contact:
    name: Ryan McCarntney
    url: https://ryan.mccartney.info/ffmpeg-docker
    email: ryan@mccartney.info
servers:
  - url: http://localhost:80/api/
paths:
...
  /file/decklink:
    post:
      description: Takes a file as an input and outputs it to a decklink card.
      tags:
        - file
      parameters:
        - in: formData
          name: filename
          type: string
          description: Filename and extension of media to playout. E.g - test.mp4
          required: true
        - in: formData
          name: cardName
          type: string
          description: The name of the BMD Decklink cards. E.g - "DeckLink SDI"
          required: true
        - in: formData
          name: font
          type: string
          description: The name of the font file to use for text overlay. Must use the
            TrueType fonts. E.g - "swansea-bold.ttf"
          required: font
        - in: formData
          name: offset
          type: number
          description: Offset for time in hours. E.g 3, -3
          required: false
        - in: formData
          name: timecode
          type: boolean
          description: Show the timecode line - true,false
          required: false
        - in: formData
          name: repeat
          type: boolean
          description: Decides whether the media loops or not
          required: false
      produces:
        - application/json
      responses:
        "200":
          description: Success
  /file/srt:
    post:
      description: SRT encode a file.
      tags:
        - file
      parameters:
        - in: formData
          name: filename
          type: string
          description: Filename and extension of media to playout. E.g - test.mp4.
            Relative to ./data/media/
          required: true
        - in: formData
          name: address
          type: string
          description: Address to direct stream towards
          required: true
        - in: formData
          name: port
          type: number
          description: Port to direct stream towards
          required: true
        - in: formData
          name: latency
          type: number
          description: SRT latency in milliseconds, default is 250ms
          required: false
        - in: formData
          name: bitrate
          type: number
          description: The bitrate of the encoded stream in kilobits per second
          required: true
        - in: formData
          name: font
          type: string
          description: The name of the font file to use for text overlay. Must use the
            TrueType fonts. E.g - "swansea-bold.ttf"
          required: false
        - in: formData
          name: offset
          type: number
          description: Offset for time in hours. E.g 3, -3
          required: false
        - in: formData
          name: timecode
          type: boolean
          description: Show the timecode line - true,false
          required: false
        - in: formData
          name: repeat
          type: boolean
          description: Decides whether the media loops or not
          required: false
      produces:
        - application/json
      responses:
        "200":
          description: Success
...
  /file/metadata:
    get:
      description: Get the metadata in a media file.
      tags:
        - files
      produces:
        - application/json
      parameters:
        - in: formData
          name: filename
          type: string
          description: The filename including extension in the `./data/media` directory
          required: false
      responses:
        "200":
          description: Success
  /file/list:
    get:
      description: Gets a list of files in the "./data/media" folder.
      tags:
        - file
      produces:
        - application/json
      parameters:
        - in: formData
          name: extension
          type: boolean
          description: Includes the file extension in the returned name
          required: false
      responses:
        "200":
          description: Success
...